requirements.txt
Django>=3.2,<3.3
djangorestframework>=3.12,<3.13
django-filter>=2.4,<2.5
Markdown>=3.3.4,<3.4